create folder structure for different modules
