+++ enemy_localization/enemySuicideDetector.py
